# docs/api/catalog-v2.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FlavorBuddy owned-app catalog API",
    "version": "2.0.0",
    "description": "Read-only search and immutable UUID documents. Obtain catalog:read tokens through /api/integrations/v1/oauth/token with resource set to the exact /api/integrations/v2/ origin/path. V1 tokens cannot access v2 and vice versa. Both versions share grants, revocation and quotas. Unknown/repeated parameters are rejected. No voting, generation, quantity inference or editorial approval."
  },
  "servers": [
    {
      "url": "https://flavorbuddy-staging-560489769953.us-central1.run.app/api/integrations/v2"
    }
  ],
  "security": [
    {
      "catalogToken": []
    }
  ],
  "paths": {
...
    "/recipes/{recipe_id}": {
      "get": {
        "operationId": "getRecipeDocument",
        "description": "Granted finalized identity only. Private/archived granted identity returns 410 with no content; unknown or ungranted identity returns 404. Stable ingredient/step IDs are scoped to the recipe UUID. Refresh availability before each cooking session.",
        "parameters": [
          {
            "name": "recipe_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Immutable document and current status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "document",
                    "status"
                  ],
                  "properties": {
                    "document": {
                      "$ref": "#/components/schemas/RecipeDocumentV1"
                    },
                    "status": {
                      "type": "object",
                      "required": [
                        "recipe_id",
                        "availability",
                        "feedback"
                      ],
                      "properties": {
                        "recipe_id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "availability": {
                          "enum": [
                            "available",
                            "unavailable"
                          ]
                        },
                        "feedback": {
                          "type": "null"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "410": {
            "description": "Unavailable; never substitute content",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "document",
                    "status"
                  ],
                  "properties": {
                    "document": {
                      "type": "null"
                    },
                    "status": {
                      "type": "object",
                      "required": [
                        "recipe_id",
                        "availability",
                        "feedback"
                      ],
                      "properties": {
                        "recipe_id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "availability": {
                          "enum": [
                            "available",
                            "unavailable"
                          ]
                        },
                        "feedback": {
                          "type": "null"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid/unsupported parameters or cursor"
          },
          "401": {
            "description": "Invalid/expired/revoked app token"
          },
          "403": {
            "description": "Missing catalog:read scope"
          },
          "404": {
            "description": "Unknown or ungranted catalog/recipe"
          },
          "429": {
            "description": "Shared app/global quota exceeded"
          },
          "503": {
            "description": "Stored recipe cannot be represented by the document contract"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "catalogToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Opaque OAuth client-credentials token; catalog:read required."
      }
    },
...
  }
}